Adding End Product Stock Adjustment, fixing active menu icon display, updating InventoryService, inventory flag, and db structure script

+++ IntrManApp.Api/Entities/ProductCheckInLine.cs
@@ -13,7 +13,7 @@ public partial class ProductCheckInLine
 
     public Guid? ProductId { get; set; }
 
-    public int TotalBatches { get; set; }
+    public decimal TotalBatches { get; set; }
 
     public decimal QuantityPerBatch { get; set; }
